@@ -1,23 +1,21 @@
 {
   "name": "@thebespokepixel/n-selector",
-  "version": "0.1.3",
+  "version": "0.2.0",
   "description": "Smart multi-gang enumerator with global, namespaced event emission for node.js.",
   "main": "index.js",
   "directories": {
     "test": "test"
   },
+  "module": "index.mjs",
   "files": [
     "index.js",
-    "lib"
+    "index.mjs"
   ],
   "scripts": {
     "test": "xo && nyc ava",
     "doc-serve": "documentation serve --watch --theme node_modules/documentation-theme-bespoke --github --config src/docs/documentation.yml --name $npm_package_name  --project-version $npm_package_version src/index.js",
     "doc-build": "documentation build --format html --output docs --theme node_modules/documentation-theme-bespoke --github --config src/docs/documentation.yml --name $npm_package_name  --project-version $npm_package_version src/index.js",
-    "readme": "compile-readme -u src/docs/example.md src/docs/readme.md > readme.md",
-    "code-climate": "nyc report --reporter=text-lcov | codeclimate-test-reporter",
-    "clean": "gulp clean",
-    "postpublish": "greenkeeper-postpublish"
+    "readme": "compile-readme -u src/docs/example.md src/docs/readme.md > readme.md"
   },
   "repository": {
     "type": "git",
@@ -34,34 +32,37 @@
   },
   "homepage": "https://github.com/MarkGriffiths/n-selector#readme",
   "copyright": {
-    "year": "2016",
+    "year": "2018",
     "owner": "Mark Griffiths @ The Bespoke Pixel"
   },
   "devDependencies": {
-    "@thebespokepixel/cordial": "^0.17.11",
-    "ava": "^0.17.0",
-    "codeclimate-test-reporter": "^0.4.0",
-    "documentation": "^4.0.0-beta15",
+    "@babel/core": "^7.0.0-beta.51",
+    "@babel/preset-env": "^7.0.0-beta.51",
+    "ava": "^0.25.0",
+    "babel-plugin-lodash": "^3.3.4",
+    "documentation": "^8.0.0",
     "documentation-theme-bespoke": "^0.3.5",
-    "greenkeeper-postpublish": "^1.0.1",
-    "gulp": "github:gulpjs/gulp#4.0",
-    "nyc": "^10.0.0",
-    "xo": "^0.17.0"
+    "gulp": "^4.0.0",
+    "gulp-better-rollup": "^3.2.1",
+    "gulp-rename": "^1.3.0",
+    "gulp-strip-comments": "^2.5.2",
+    "nyc": "^12.0.2",
+    "rollup-plugin-babel": "^4.0.0-beta.5",
+    "xo": "^0.21.1"
   },
   "xo": {
     "semicolon": false,
     "esnext": true,
     "ignores": [
       "index.js",
-      "lib/**",
-      "docs/**",
-      "bin/**",
-      "test/coverage/**"
+      "index.mjs",
+      "test/coverage/**",
+      "docs"
     ]
   },
   "dependencies": {
-    "lodash": "^4.16.6",
-    "sparkles": "^1.0.0"
+    "lodash": "^4.17.10",
+    "sparkles": "^1.0.1"
   },
   "badges": {
     "github": "MarkGriffiths",
@@ -116,9 +117,7 @@
       ]
     ]
   },
-  "buildNumber": 0,
   "engines": {
-    "node": ">=4.0 <9.0"
-  },
-  "jsnext:main": "lib/index-es.js"
+    "node": ">=8.0"
+  }
 }
