@@ -1,48 +1,48 @@
-/*
- * Client Gulp Template Tasks
- * (Work in progress)
- */
- /* eslint import/no-extraneous-dependencies: 0 */
-
 const gulp = require('gulp')
-const cordial = require('@thebespokepixel/cordial')()
+const rename = require('gulp-rename')
+const strip = require('gulp-strip-comments')
+const rollup = require('gulp-better-rollup')
+const babel = require('rollup-plugin-babel')
+const lodash = require('babel-plugin-lodash')
 
-// transpilation/formatting
-gulp.task('bundle', cordial.macro({
-	source: 'src/index.js'
-}).bundle())
+const external = ['lodash/isInteger', 'lodash/maxBy', 'lodash/filter', 'lodash/head', 'lodash/map', 'sparkles', 'crypto']
 
-gulp.task('master', cordial.macro({
-	master: true,
-	source: 'src/index.js'
-}).bundle())
+const babelConfig = {
+	plugins: [lodash],
+	presets: [
+		['@babel/preset-env', {
+			modules: false,
+			targets: {
+				node: '8.0.0'
+			}
+		}]
+	],
+	exclude: 'node_modules/**'
+}
 
-// Clean
-gulp.task('clean', cordial.shell({
-	source: ['npm-debug.*', './nyc_output', './test/coverage']
-}).trash())
+gulp.task('cjs', () =>
+	gulp.src('src/index.js')
+		.pipe(rollup({
+			external,
+			plugins: [babel(babelConfig)]
+		}, {
+			format: 'cjs'
+		}))
+		.pipe(strip())
+		.pipe(gulp.dest('.'))
+)
 
-// Docs
-gulp.task('docs', cordial.shell({
-	source: 'npm run doc-build'
-}).job())
+gulp.task('es6', () =>
+	gulp.src('src/index.js')
+		.pipe(rollup({
+			external,
+			plugins: [babel(babelConfig)]
+		}, {
+			format: 'es'
+		}))
+		.pipe(strip())
+		.pipe(rename('index.mjs'))
+		.pipe(gulp.dest('.'))
+)
 
-// ReadMe
-gulp.task('readme', cordial.shell({
-	source: 'npm run readme'
-}).job())
-
-// Tests
-gulp.task('ava', cordial.test().ava(['test/*.js']))
-gulp.task('xo', cordial.test().xo(['src/lib/*.js']))
-gulp.task('test', gulp.parallel('xo', 'ava'))
-
-// Default
-gulp.task('default', gulp.series('bump', 'bundle'))
-
-// Hooks
-gulp.task('start-release', gulp.series('reset', 'clean', 'master', 'readme'))
-gulp.task('post-flow-release-start', gulp.series('start-release', 'version-release', 'docs', 'commit'))
-
-// Default
-gulp.task('default', gulp.series('bump', 'clean', gulp.parallel('docs', 'bundle', 'readme')))
+gulp.task('default', gulp.series('cjs', 'es6'))
